Issue #SB-26129 bug fix: App icon is stored as google drive url link on reading content in dock.

+++ platform-modules/import-manager/src/main/scala/org/sunbird/object/importer/ImportManager.scala
@@ -15,7 +15,6 @@ import org.sunbird.common.exception.{ClientException, ServerException}
 import org.sunbird.graph.OntologyEngineContext
 import org.sunbird.graph.common.Identifier
 import org.sunbird.graph.utils.ScalaJsonUtils
-import org.sunbird.telemetry.logger.TelemetryManager
 import org.sunbird.telemetry.util.LogTelemetryEventUtil
 import org.sunbird.url.common.URLErrorCodes
 import org.sunbird.url.util.{GoogleDriveUrlUtil, HTTPUrlUtil}
@@ -76,12 +75,16 @@ class ImportManager(config: ImportConfig) {
 					sourceMetadata
 				} else reqMetadata
 				val appIcon = finalMetadata.getOrDefault("appIcon","").asInstanceOf[String]
-				if(appIcon != null && appIcon.nonEmpty)
+				if(appIcon != null && appIcon.nonEmpty && !appIcon.contains(ss.getContainerName()))
 				{
 					try {
 						if (!appIconMap.containsKey(appIcon)) {
-							val appIconFolder = CONTENT_FOLDER + File.separator + Slug.makeSlug(finalMetadata.getOrDefault("identifier", "").asInstanceOf[String], true) + File.separator + "assets"
-							val appIconFile = downloadAppIconFile(finalMetadata.getOrDefault("identifier", "").asInstanceOf[String], appIcon)
+							val identifier = {
+								if(finalMetadata.containsKey("identifier")) finalMetadata.getOrDefault("identifier", "").asInstanceOf[String]
+								else finalMetadata.getOrDefault("collectionId", "").asInstanceOf[String]
+							}
+							val appIconFolder = CONTENT_FOLDER + File.separator + Slug.makeSlug(identifier, true) + File.separator + "assets"
+							val appIconFile = downloadAppIconFile(identifier, appIcon)
 							val appIconCloudUrl = ss.uploadFile(appIconFolder, appIconFile, Option(false))(1)
 							try {
 								if(appIconFile.exists()) FileUtils.deleteDirectory(appIconFile.getParentFile.getParentFile)
